Move tests to Primitives
