+++ System.IO.FileSystem/tests/System.IO.FileSystem.Tests/Primitives/FileAccessTests.cs
@@ -3,7 +3,7 @@
 
 using nanoFramework.TestFramework;
 
-namespace System.IO.FileSystem.Tests
+namespace System.IO.FileSystem.Tests.Primitives
 {
     [TestClass]
     public static class FileAccessTests
