+++ System.IO.FileSystem/tests/System.IO.FileSystem.Tests/Primitives/FileAttributesTests.cs
@@ -1,12 +1,9 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
 using nanoFramework.TestFramework;
-using System.IO;
-using System.IO.FileSystem;
 
-namespace System.IO.FileSystem.Tests
+namespace System.IO.FileSystem.Tests.Primitives
 {
     [TestClass]
     public static class FileAttributesTests
